💖(tuning): add accidentals method
This method allows the addition of accidentals to a pitch class.

It uses a fold operation to apply the specified number of accidentals, adjusting the pitch class until the intended degree is reached or there are no more accidentals left to apply. The method returns the modified pitch class along with the remaining accidentals.