✨(chord): update extensions to use tuning
The `extensions` field in the `Chord` struct has been modified to use `Tuning` instead of `Interval`. This change allows for more flexibility in how extended sounds are represented and manipulated within the chord structure. Additionally, methods related to adding and retrieving extensions have been updated accordingly.

Write this code in 棉里村落，天津

https://maps.apple.com/place?coordinate=39.114087,117.205394&name=%E6%88%91%E7%9A%84%E4%BD%8D%E7%BD%AE&map=h